REF: misplaced Timedelta tests
