+++ pandas/tests/indexes/timedeltas/test_scalar_compat.py
@@ -121,3 +121,15 @@ class TestVectorizedTimedelta:
             tm.assert_index_equal(r1, s1)
             r2 = t2.round(freq)
             tm.assert_index_equal(r2, s2)
+
+    def test_components(self):
+        rng = timedelta_range("1 days, 10:11:12", periods=2, freq="s")
+        rng.components
+
+        # with nat
+        s = Series(rng)
+        s[1] = np.nan
+
+        result = s.dt.components
+        assert not result.iloc[0].isna().all()
+        assert result.iloc[1].isna().all()
